fix: improve user prompts and script flow in install.sh

+++ install.sh
@@ -1,7 +1,25 @@
+#!/bin/bash
+
+read -p "Do you want to run this script? (y/n): " confirm
+
+if [[ "$confirm" != "y" && "$confirm" != "Y" ]]; then
+    echo "Exiting..."
+    exit 0
+fi
+
+echo "Starting..."
+
 # if sudo is not installed
 # sudo apt install sudo
 # use this if usermod did not work
 sudo usermod -aG input,video,plugdev $user
+
+read -p "Do you want install anything? (y/n): " confirm
+
+if [[ "$confirm" != "y" && "$confirm" != "Y" ]]; then
+    echo "Exiting..."
+    exit 0
+fi
 
 echo "Installing wget, gcc, and make temporarily..."
 sudo apt install wget gcc make libpam0g-dev
